pyproject.toml
[project]
name = "mcp-geom"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "ase>=3.28.0",
    "mcp>=1.26.0",
    "numpy>=2.4.4",
    "openai>=2.30.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
